quad_gantry_level: Retry based off of z adjust
determine wether or not to retry
based off the largest z stepper adjustment.

instead of triggering based off of the range of hte probed z heights.
(difference between the highest and lowest)

This should be more intuitive.

It is also intended to match the interface of a similar feature to be
added to z_tilt_adjust and z_tilt_adjust has an edge case where using
the range of probed points is awkward. (3 probe points, 2 z steppers)

Signed-off-by: John "Fess" Fessenden <fess@fess.org>